+++ docs/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implement a way to switch off each collector (OSIDB-2884)
 - Generate Jira tracker "components" field (OSIDB-2988)
 - Rudimentary API request logging (OSIDB-2514)
+- Add query param to force creation of Jira task for old flaws on update (OSIDB-2882)
 
 ### Changed
 - Update the SLA policy